fixing tests to support systems with different `Int`

+++ test/iterators.jl
@@ -25,17 +25,17 @@ end
 # unzip
 let z = zip(1:2)
     @test unzip(z) == 1:2
-    @test eltype(unzip(z)) == UnitRange{Int64}
+    @test eltype(unzip(z)) == UnitRange{Int}
 end
 
 let z = zip(1:2, 3:4)
     @test collect(unzip(z)) == [(1,2), (3,4)]
-    @test eltype(unzip(z)) == UnitRange{Int64}
+    @test eltype(unzip(z)) == UnitRange{Int}
 end
 
 let z = zip(1:2, 3:4, 5:6)
     @test collect(unzip(z)) == [(1,2), (3,4), (5,6)]
-    @test eltype(unzip(z)) == UnitRange{Int64}
+    @test eltype(unzip(z)) == UnitRange{Int}
 end
 
 @test eltype(Iterators.filter(isodd, 1:5)) == Int
